backend_python/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
colorama==0.4.4
distlib==0.3.2
Django==3.2.7
django-cors-headers==3.8.0
django-nose==1.4.7
djangorestframework==3.12.4
filelock==3.0.12
nose==1.3.7
packaging==21.0
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tox==3.24.3
virtualenv==20.7.2
